Fixed assert failure on call to opCast(?&out) with a non-variable expression
git-svn-id: http://svn.code.sf.net/p/angelscript/code/trunk@2509 404ce1b2-830e-0410-a2e2-b09542c77caf

+++ sdk/tests/test_feature/source/test_shared.cpp
@@ -31,7 +31,7 @@ bool Test()
 		r = mod->Build();
 		if (r < 0) TEST_FAILED;
 
-		size_t currAlloc = GetAllocedMem();
+		int currAlloc = GetAllocedMem();
 
 		// discard `Module2`
 		mod->Discard();
